Rename reducers to match actions.
