+++ src/js/reducers/hintsReducer.js
