+++ src/js/reducers/index.js
@@ -3,13 +3,13 @@ var combineReducers = require('redux-immutable').combineReducers;
 
 // order matters here
 module.exports = combineReducers({
-  scene: require('./scene'),
-  vega: require('./vega'),
-  inspector: require('./inspector'),
-  marks: require('./marks'),
-  scales: require('./scales'),
-  signals: require('./signals'),
-  walkthrough: require('./walkthrough'),
-  hints: require('./hints'),
-  pipelines: require('./pipelines')
+  scene: require('./sceneReducer'),
+  vega: require('./vegaReducer'),
+  inspector: require('./inspectorReducer'),
+  marks: require('./marksReducer'),
+  scales: require('./scalesReducer'),
+  signals: require('./signalsReducer'),
+  walkthrough: require('./walkthroughReducer'),
+  hints: require('./hintsReducer'),
+  pipelines: require('./pipelinesReducer')
 });
