+++ src/js/reducers/inspectorReducer.js
