+++ src/js/reducers/pipelinesReducer.js
