+++ src/js/reducers/scalesReducer.js
