+++ src/js/reducers/signalsReducer.js
