+++ src/js/reducers/signalsReducer.test.js
@@ -4,7 +4,7 @@ var expect = require('chai').expect;
 var Immutable = require('immutable');
 
 var actions = require('../actions/Names');
-var signalsReducer = require('./signals');
+var signalsReducer = require('./signalsReducer');
 var markActions = require('../actions/markActions');
 var createScene = require('../actions/sceneActions').createScene;
 var signalActions = require('../actions/signalActions');
