+++ src/js/reducers/vegaReducer.js
