+++ src/js/reducers/walkthroughReducer.js
